Order chains consistently
Relying on the hash's "order" results in these lines of the
iptables.rules file swapping randomly:

  :FORWARD DROP [0:0]
  :INPUT DROP [0:0]
  :OUTPUT ACCEPT [0:0]

This causes deployments to unnecessarily rewrite the iptables.rules
file (which in turn causes iptables to unnecessarily be flushed and
restored, and clutters up other moonshine-related output).

This commit orders the chains consistently to reduce unnecessary
work, risk, and output.